+++ src/main/java/kz/brk/imports/model/StcTbl05DhdRlzPrdUsl.java
@@ -1,6 +1,7 @@
 package kz.brk.imports.model;
 import lombok.*;
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.util.Date;
 @Builder
 @Entity
@@ -16,12 +17,12 @@ public class StcTbl05DhdRlzPrdUsl {
   private int id;
   private String idSubclass;
   private int god;
-  private int fld015DhdRlzc;
-  private int fld016DhdFin;
-  private int fld017DvdAkcVzng;
-  private int fld018PrDhd;
-  private int fld019DhdKursRznc;
-  private int fld020DhdVybAkt;
+  private BigDecimal fld015DhdRlzc;
+  private BigDecimal fld016DhdFin;
+  private BigDecimal fld017DvdAkcVzng;
+  private BigDecimal fld018PrDhd;
+  private BigDecimal fld019DhdKursRznc;
+  private BigDecimal fld020DhdVybAkt;
   @Builder.Default
   private int relevance = 1;
   @Builder.Default
